#3: Solved routing error, need to refactor later
