+++ components/index.js
@@ -3,5 +3,6 @@ import Link from "./link";
 import Route from "./route";
 import ColorPad from "./ColorPad";
 import PrecentageBar from "./PrecentageBar";
+import PlayerController from "./PlayerController";
 
-export { ColorPad, Effect, Link, Route, PrecentageBar };
+export { ColorPad, Effect, Link, Route, PrecentageBar, PlayerController };
