+++ components/route/index.js
@@ -1,14 +1,13 @@
 import { useContext } from "react";
 import { AppContext } from "../../contexts";
 
-const Route = ({ path, component }) => {
-  const [appState, setAppState] = useContext(AppContext);
-  if (appState.path === path) {
-    return component({
-      appState: appState,
-      setAppState: setAppState,
-    });
-  }
+import { View } from "react-native";
+const Route = ({ path, component, currentPath, setCurrentPath }) => {
+  return (
+    <View name={path}>
+      {currentPath == path && component({ setCurrentPath: setCurrentPath })}
+    </View>
+  );
 };
 
 export default Route;
