+++ views/Game/index.tsx
@@ -1,13 +1,37 @@
 import { View, Text,SafeAreaView, StyleSheet } from "react-native";
-import { generateColor, Screen, Player } from "../../constants";
+import { Screen, Player } from "../../constants";
 import { useEffect, useState, useRef } from "react";
-import { MathHelper } from '../../helpers';
 import * as Animatable from 'react-native-animatable';
-import {failSound,stopSound,startSound,normalSound,clockSound,pressSound} from "../../helpers/SoundPlayer";
-import {PrecentageBar, ColorPad} from "../../components";
+import {startSound,normalSound,clockSound} from "../../helpers/SoundPlayer";
+import {PlayerController} from "../../components";
 
-
-const CountDown = ({active,countDownRef, countDown}) => {
+const CountDown = ({active,countDown}) => {
+  const countDownRef =  useRef<Animatable.View & View>(null);
+  const flipEven = {
+    0: {
+      transform: [{rotateZ: "0deg"}],
+    },
+    1: {
+      transform: [{rotateZ: "180deg"}],
+    },
+  };
+  const flipOdd = {
+    0: {
+      transform: [{rotateZ: "180deg"}],
+    },
+    1: {
+      transform: [{rotateZ: "360deg"}],
+    },
+  };
+  useEffect(() => {
+    setTimeout(()=> {
+      if (countDown == 0) {
+        countDownRef.current && countDownRef.current.flipOutY();
+      } else {
+        countDownRef.current.animate(countDown % 2 ? flipEven : flipOdd);
+      }
+    }, 1000);
+  }, [countDown]);
   return (
     <View style={{display: !active ? "flex" : "none",position: "absolute",width:"100%",height:"100%",alignItems:"center",justifyContent:"center"}}>
     <Animatable.View ref={countDownRef} animation="flipInY" iterationCount={1} style={{width: Screen.Width * 0.9,height:Screen.Width * 0.9,backgroundColor:"black",borderRadius: Screen.Width * 0.9,display: "flex",alignItems:"center",justifyContent:"center"}}>
@@ -17,106 +41,13 @@ const CountDown = ({active,countDownRef, countDown}) => {
   )
 }
 
-const PlayerController = (props) => {
-  const { player,active,setGameState,gameState } = props;
-  const [points,setPoints] = useState<number>(0);
-  const [powerMeter,setPowerMeter] = useState<number>(100);
-  const [time,setTime] = useState<number>(MathHelper.randomIntFromInterval(2, 10));
-  const [color,setColor] = useState<string>(generateColor());
-  const animatableRef = useRef<Animatable.View & View>(null); 
 
 
-  const onClick = (clickColor:string,player:string)=> {
-    
-    if(gameState.blocked == player){
-      setPowerMeter(0);
-      setPoints(0);
-      failSound.play();
-      const backgroundColor = animatableRef.current.props.style.backgroundColor
-      animatableRef.current && animatableRef.current.animate({0:{transform: [{scale: 1},{rotateZ: "0deg"}]},0.33:{transform: [{scale: 1.5},{rotateZ: "10deg"}]},0.66:{transform: [{scale: 1.5},{rotateZ: "-10deg"}]},1:{transform: [{scale: 1},{rotateZ: "0deg"}]}})
-    }
-    if(clickColor == color) {
-      setPoints((prevPoints) => (prevPoints + 1));
-      if(powerMeter < 100){
-        setPowerMeter(powerMeter + 10);
-      }
-      pressSound.play();
-    } else if(clickColor != color) {
-      if(points > 0){
-        setPoints((prevPoints) => (prevPoints - 1));
-      }
-      pressSound.play();
-    }
-  };
-
-  useEffect(() => {
-    let intervall = setInterval(()=> {
-      if (time == 0) {
-        setColor(generateColor());
-        setTime(MathHelper.randomIntFromInterval(1, 5));
-      } else {
-        setTime(time - 1);
-      }
-    }, 1000);
-    return () => {
-      clearInterval(intervall);
-    };
-  }, [time]);
-
-  const styles = StyleSheet.create({
-    wrapper: {
-      display: "flex",
-      height: "100%",
-      width: "100%",
-      flexDirection: player == Player.One ? "column" : "column-reverse",
-    },
-    innerWrapper: {
-      display: "flex",
-      flexDirection: "row",
-      justifyContent: "space-between",
-      padding:15
-    },
-    precentageBarWrapper:{height: 80,display: "flex",justifyContent:"center"},
-    pointsWrapper: {display: "flex",flexDirection: player == Player.One ? "column-reverse" : "column",alignItems:"center",justifyContent: player == Player.One ? "flex-end" : "flex-start"}
-  });
-
-  const blockPlayer = () => {
-    stopSound.play();
-    normalSound.setMute(true);
-    setPowerMeter(0);
-    setGameState({...gameState,blocked: player === Player.One ? Player.Two : Player.One})
-  }
-
-  return (
-        <View style={styles.wrapper}>
-          <View style={styles.innerWrapper}>
-            <ColorPad gameState={gameState} player={player} active={active} _onClick={onClick} />
-            <ColorPad gameState={gameState} player={player} active={active} _onClick={onClick} />
-            <ColorPad gameState={gameState} player={player} active={active} _onClick={onClick} />
-            <ColorPad gameState={gameState} player={player} active={active} _onClick={onClick} />
-          </View>
-          <View style={styles.precentageBarWrapper}>
-            <PrecentageBar gameState={gameState} player={player} onPressStop={blockPlayer} value={powerMeter} color={color}/>
-            </View>
-            <View style={styles.pointsWrapper}>
-              <View style={{height:150}}/>
-              <Animatable.View ref={animatableRef} duration={1500} style={{width: Screen.Width * 0.2,height: Screen.Width * 0.2,borderRadius: 50,backgroundColor:color,display: "flex",alignItems:"center",justifyContent: "center"}}>
-              <Text style={{fontSize: 50,color: "white"}}>{points}</Text>
-            </Animatable.View>
-          </View>
-        </View>
-  );
-};
-
-
-
-export default function Game(props) {
-  const { appState,setAppState } = props;
-  const [gameTime,setGameTime] = useState<number>(120);
-  const [countDown,setCountDown] = useState<number>(3);
+export default function Game({setCurrentPath}:any) {
   const [active,setActive] = useState<boolean>(false);
-  const countDownRef =  useRef<Animatable.View & View>(null);
+  const [countDown,setCountDown] = useState<number>(3);
   const [gameState,setGameState] = useState<any>({mode: "oneColor",blocked: null});
+  const [gameTime,setGameTime] = useState<number>(0);
 
   useEffect(() => {
     if(clockSound.ready){
@@ -138,34 +69,31 @@ export default function Game(props) {
   },[normalSound.ready]);
   
   useEffect(() => {
-    let intervall = setInterval(()=> {
-      if (gameTime == 0) {
-        setAppState({...appState,path: "score"})
-      } else {
-        setGameTime(gameTime - 1);
-      }
-    }, 1000);
-    return () => {
-      clearInterval(intervall);
-    };
-  }, [gameTime]);
-
-  useEffect(() => {
     setTimeout(()=> {
       if(countDown == 1){
         setGameState({...gameState,mode: ""})
       }
       if (countDown == 0) {
-        countDownRef.current && countDownRef.current.flipOutY();
         setTimeout(() => {
-          setActive(true);
-        },1000);
+        setGameTime(60);
+        setActive(true);
+      },1000);
       } else {
         setCountDown(countDown - 1);
-        countDownRef.current.animate(countDown % 2 ? flipEven : flipOdd);
       }
     }, 1000);
   }, [countDown]);
+  useEffect(() => {
+    if(countDown == 0){
+      setTimeout(()=> {
+        if (gameTime == 0) {
+          setCurrentPath("score");
+        } else {
+          setGameTime(gameTime - 1);
+        }
+      }, 1000);
+    }
+  }, [gameTime]);
 
   useEffect(() => {
     if(gameState.blocked){
@@ -178,23 +106,6 @@ export default function Game(props) {
       },5000);
     }
   },[gameState]);
-
-  const flipEven = {
-    0: {
-      transform: [{rotateZ: "0deg"}],
-    },
-    1: {
-      transform: [{rotateZ: "180deg"}],
-    },
-  };
-  const flipOdd = {
-    0: {
-      transform: [{rotateZ: "180deg"}],
-    },
-    1: {
-      transform: [{rotateZ: "360deg"}],
-    },
-  };
 
   const styles = StyleSheet.create({
     wrapper: { width: "100%", height: "100%" },
@@ -209,7 +120,7 @@ export default function Game(props) {
     middleWrapperText: {fontSize: Screen.Width * 0.2, fontWeight: "100",color: "black"},
     playerWrapper: { width: "100%", height: "50%" }
   });
-  return (
+  return ( 
     <SafeAreaView>
     <View style={styles.wrapper}>
       <View
@@ -227,7 +138,7 @@ export default function Game(props) {
         </View>
       </View>
     </View>
-    <CountDown countDownRef={countDownRef} active={active} countDown={countDown}/>
+    <CountDown active={active} countDown={countDown}/>
     </SafeAreaView>
   );
 }
